Added upgrade handling for 'My SubWorkflows' and outdated Package subworkflows along with namespace changes required to support these upgrades.
Summary:

- Subworkflows from 'My SubWorkflows' are automatically upgraded when vistrails is loaded.
- Package subworkflows are no longer editable from the GUI by clicking "Edit SubWorkflow".  This is to prevent users from modifying the underlying package.  Package creators can directly open the Subworkflow's xml file in VisTrails to make modifications for use in a new version of the package.  Users wanting to customize a package subworkflow for their own use must first import it into their 'My SubWorkflows' (where it becomes disassociated from the package).
- Subworkflows from an outdated Package are automatically upgraded when vistrails is loaded, but the original package is not modified; instead a temporary copy of the upgraded subworkflow is kept, and if the upgraded subworkflow is used in a vistrail, a non-editable copy of the upgraded subworkflow is stored with the vistrail.  (The non-editable status is set by an '__abstraction_descriptor_info__' annotation, which stores the descriptor info of the original package subworkflow [for later replacement if desired].  This annotation is removed if the upgraded subworkflow is imported into 'My SubWorkflows').
- A vistrail can now have subworkflows with the same name from different packages (previously you could not have subworkflows with the same name in a single vistrail, even if they were from different packages or namespaces).

Details:

core/modules/abstraction.py

  initialize():
    - Modified to handle upgrades for 'My SubWorkflows' package

core/modules/module_registry.py

  ModuleRegistry:
    add_subworkflow():
      - Modified to handle upgrades for outdated package subworkflows

core/vistrail/abstraction.py

  Abstraction
    is_latest_version():
      update_version_status():
        - Modified to disconnect version updates on subworkflow modules when the vistrail is closed

core/vistrail/controller.py

  VistrailController
    __init__():
      - Modified to use new flush_pipeline_cache() function (see below)

    flush_pipeline_cache():
      - Added function

    close_vistrail():
      - Modified to not unload abstractions in an abstraction (otherwise, editing an abstraction and closing it unloads abstractions still needed by the parent vistrail).

    get_abstraction_desc():
      - Modified to accept package name as a parameter

    parse_abstraction_name():
      - Modified to handle all types of abstraction filenames, including the new format which uses uuid namespace in the filename.
      - Maintains backwards compatibility to only get the abstraction name, but can also pass a flag to get all parts of the parse (path, prefix, abstraction name, namespace, suffix)

    add_abstraction_to_registry():
      - Modified to ensure module_version is a str in call to registry.auto_add_module()

    load_abstraction():
      - Modified to read/set '__abstraction_origin_uuid__' and '__abstraction_uuid' annotations, allowing for backwards compatibility
      - Modified to pass local abstractions package to newly modified get_abstraction_desc() (see above)

    unload_abstractions():
      - Modified to not unload manually upgraded package abstractions (if an older version from the same package is somehow loaded) because it would also remove the new version from the module registry. (This case may not even be possible - possibly worth looking into later).

    update_abstraction():
      - Commented out old function

    save_abstraction():
      - Modified to read/set '__abstraction_origin_uuid__' and '__abstraction_uuid' annotations, allowing for backwards compatibility

    unload_abstraction():
      - Commented out old function [not to be confused with unload_abstractions()]

    import_abstraction():
      - Modified to accept package name as a parameter
      - Modified to strip '__abstraction_descriptor_info__' annotation on import

    export_abstraction():
      - Modified to accept package name as a parameter

    find_abstractions():
      - Modified to accept flag parameter that includes upgraded package abstractions

    check_abstraction():
      - Modified to use (name,namespace) tuple for lookup (see ensure_abstractions_loaded() below)

    ensure_abstractions_loaded():
      - Modified to build lookup from (name,namespace) tuples rather than just 'name'.

    handle_invalid_pipeline():
      - Modified to accept flag parameter that forces upgrades to be immediate, even if upgradeDelay is set to True

    write_vistrail():
      - Modified to use updated version of find_abstractions() to also get upgraded package abstractions
      - Modified to handle namespace updating when saving to a new file, or when saving upgraded package abstractions with the vistrail
      - Modified to handle unloading abstractions from old namespace, loading of abstractions from new namespace when saving to a new file, or when saving upgraded package abstractions with the vistrail

      make_abstraction_path_unique():
        - Added function (to improve readibility) that creates the unique abstraction path and file for saving into the vistrail

core/vistrail/pipeline.py

  Pipeline
    validate():
      - Modified abstraction upgrade checks to function properly if types don't match (str vs. int)

core/vistrail/vistrail.py

  Vistrail
    set_annotation():
      - Modified to delete an annotation if 'value' is None or contains only spaces

    get_latest_version():
      - Modified to fix upgrade return to be long and not str

gui/builder_window.py

  QBuilderWindow
    editAbstraction():
      - Modified to prevent editing of a package subworkflow or an upgraded package subworkflow (stored or not)

gui/vistrail_controller.py

  VistrailController
    import_abstractions():
      - Modified to include package name in call to import_abstraction

    export_abstractions():
      - Modified to include package name in call to export_abstraction

+++ vistrails/core/vistrail/pipeline.py
@@ -803,7 +803,7 @@ class Pipeline(DBWorkflow):
                 if module.is_abstraction():
                     try:
                         desc = module.module_descriptor
-                        if module.internal_version != desc.version:
+                        if long(module.internal_version) != long(desc.version):
                             exceptions.add(MissingModuleVersion(desc.package, desc.name, desc.namespace, desc.version, desc.package_version, module.id))
                     except:
                         pass
